Update cardinal-principles.md
We have decided to remove the last point of the cardinal principles for the reason that it is superfluous in the face of all other points in the cardinal principles, but not due to any disagreement with its content. The cardinal principles are meant to be the most reduced position necessary for membership in RAIM, and therefore it helps to remain as concise as possible in its explanation. Further, we will be adding the line "We reject postmodern, liberal-utopian and technocratic "alternatives" to scientific socialism, and uphold the legacy of Marxism, Leninism and Maoism."